+++ src/App.jsx
@@ -1,20 +1,17 @@
 import { BrowserRouter as Router, Route, Routes } from 'react-router-dom';
-import './App.css';
-
-// src/pages からコンポーネントをインポート
-import Home from './pages/home';
-import About from './pages/about';
-import Text from './pages/Text';  // 新しいページ（Text）をインポート
+import Home from './pages/Home/Home';
+import Text from './pages/Text/Text';
+import About from './pages/About/About';
+import './App.css'; // アプリ全体のスタイル
 
 function App() {
   return (
     <Router>
       <div>
-        {/* ルートとコンポーネントの設定 */}
         <Routes>
           <Route path="/" element={<Home />} />
+          <Route path="/text" element={<Text />} />
           <Route path="/about" element={<About />} />
-          <Route path="/text" element={<Text />} /> {/* 新しいページのルート */}
         </Routes>
       </div>
     </Router>
