+++ src/pages/Text/Text.jsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 import { Link } from 'react-router-dom';
-import './text.css'; // text.css ファイルをインポート
+import './text.css';
 
 const Text = () => {
   const [text, setText] = useState('');
@@ -11,10 +11,8 @@ const Text = () => {
 
   return (
     <div className="text-container">
-      {/* アプリ名 */}
       <h1 className="app-name">Reader me</h1>
 
-      {/* 文章入力エリア */}
       <div className="text-area-container">
         <textarea
           className="text-area"
@@ -24,11 +22,10 @@ const Text = () => {
         ></textarea>
       </div>
 
-      {/* Aboutページに遷移するボタン */}
-      <Link to="/about">
+      {/* 文字数と文章そのものをstateとして渡す */}
+      <Link to="/about" state={{ textLength: text.length, textContent: text }}>
         <button className="navigate-btn">Go to About</button>
       </Link>
-
     </div>
   );
 };
